Read func moved to a separate file
